+++ src/debug.py
@@ -3,4 +3,4 @@
 ###############################################
 
 def print(*args,**kwargs):
-	__import__('builtins').print(f"{__import__('os').path.relpath(__import__('inspect').stack()[1].filename,__import__('os').path.dirname(__import__('os').path.abspath(__file__)))}:{__import__('inspect').stack()[1].lineno}:{__import__('inspect').stack()[1].function}", " ### ",*args,**kwargs)
+	__import__('builtins').print(f"{__import__('os').path.relpath(__import__('inspect').stack()[1].filename,__import__('os').path.dirname(__import__('os').path.abspath(__file__)))}:{__import__('inspect').stack()[1].lineno}:{__import__('inspect').stack()[1].function}", " --- ",*args,**kwargs)